Finished adding delete functionality to front and back end

+++ app/controllers/paintings_controller.rb
@@ -5,7 +5,7 @@ class PaintingsController < ApplicationController
     end
 
     def create
-        if session[:user_id]
+        if find_user
             user = User.find(session[:user_id])
             painting = user.paintings.create!(paint_params)
             render json: painting, status: :created
@@ -15,22 +15,42 @@ class PaintingsController < ApplicationController
     end
 
     def update
-        if session[:user_id]
+        if find_user
+            user = User.find(find_user)
             painting = Painting.find(find_painting)
-            painting.update(paint_params)
-            render json: painting, status: :accepted
+
+            if painting.user_id === user.id
+                painting.update(paint_params)
+                render json: painting, status: :accepted
+            else
+                render json:  {"errors": ["Not authorized"], status: :unauthorized}
+            end
         else
             render json:  {"errors": ["Not authorized"], status: :unauthorized}
         end
     end
 
     def destroy
-        painting = Painting.find(paint_params)
-        painting.destroy
-        head :no_content
+        if find_user
+            user = User.find(find_user)
+            painting = Painting.find(find_painting)
+
+            if painting.user_id === user.id
+                painting.destroy
+                head :no_content
+            else
+                render json:  {"errors": ["Not authorized"], status: :unauthorized}
+            end
+        else
+            render json:  {"errors": ["Not authorized"], status: :unauthorized}
+        end
     end
 
     private
+
+    def find_user
+        session[:user_id]
+    end
 
     def paint_params
         params.permit(:name, :bio, :img_url, :year, :museum_id)
